handle individual entity's describeSObject error in integration test
An entity's describeSObject occassionally fails due to a bug such as an incompatible change introduced by an entity owner. This is not an issue with data loader or WSC. It is an issue with the entity, which gets fixed over time. As such, the integration test should catch and record the entity with the issue but not report it as an error in the DL integration test results.