avoid app freeze in macOS when saving field mappings 2nd time
SWT FileDialog widget has a bug in macOS which causes DL to freeze if the user tries to save field mappings twice on macOS. Solution is to remove the code lines that try to set directory and file when the dialog launches. FileDIalog continues to show the list of files in a directory that user had navigated to in the previous result, which is an acceptable behavioor.