fix the warning message by specifying a higher value for cell width.
Following warning is logged when OAuth browser window is displayed by DL:

Warning: Expected min height of view: (<NSPopoverTouchBarItemButton: 0x7fc1a9255f10>) to be less than or equal to 30 but got a height of 32.000000.

The code change avoids the display of this warning which causes unnecessary concern for the users.