fix issue with creating desktop shortcut link
Desktop shortcut link is created in %USERPROFILE%\Desktop folder, which is the case for most desktops. However, sometimes Desktop folder is configured in a different location, which results in the missing desktop shortcut link. Fix is to look up the registry to figure out the location of Desktop folder and create the link.