fix and Additional tests for date-only field processing
fix and Additional tests for date-only field processing.

Fix - set hour, minute, second, and millisec to 0 only if datefield.usegmt is not set to True (if set to true, revert to old behavior, i.e. do not set time to 12AM in local time).

Also, do not increment day by 1. Just add the offset between GMT and local timezone if local timezone is ahead of GMT.