Use the charset specified in file.encoding system property for CSV read and write
Explicitly use the value of "file.encoding" system property for charset when reading from or writing to a CSV file instead of relying on JVM to set the default charset to the value specified in "file.encoding" system property.

Add more debugging around selection of the charset.

Use the same code to select the charset for both CSV read and CSV write.

This change is based on issue #502 .