remove ui package dependency from OAuthBrowserLoginRunner
A class in util package such as OAuthBrowserLoginRunner should not depend on a class in UI package because util is used by batch mode and batch mode does not load SWT packages, on which UI package depends.

Fixed the issue by moving processSuccessfulLogin() static method from OAuthFlow class in UI package to OAuthBrowserLoginRunner class in util.