support single read from data source for small data uploads
Introduce read-only property, process.bulk.cacheDataFromDao, that when set to true (default false) will cache the uploaded data from the data source and use it to process results instead of re-uploading data from the data source. Setting this property to true will result in the code using more heap to cache the data and may result in OutOfMemoryError for large data uploads.