fix for exception "INVALID_SOAP_HEADER' exceptionMessage='You can't use KeepAccountTeam to change a Contact's owner."
fix for W-11635112 - exception "INVALID_SOAP_HEADER' exceptionMessage='You can't use KeepAccountTeam to change a Contact's owner."

Introduction of "Keep Account Team" (PR #552) has introduced the issue of KeepAccountTeam being set for insert, delete, upsert, and update operations on all sObjects, including Account. This results in the exception such as the one captured in the Subject line. The fix is to set KeepAccountTeam option of OwnerChangeOptions SOAP header only if the operation is on Account object and clear these headers if the operation is on some other object.