add BOM tests for UTF-16BE and UTF-16LE
Read-only property "dataAccess.readCharset" to specify an encoding for the CSV file to be uploaded in case it is different from the system property "file.encoding".

Change to pom.xml to be able to copy csv files in non-UTF8 encoding to target directory.

Tests for UTF16BE and UTF16LE encoded CSV files where header field names are in double quotes.