Fix W-12572271 - "Finish" button of Data Loader upsert is grayed out at first time after login
Finish button is enabled when all of Wizard steps are marked as "Completed". If an sObject does not have a modifiable lookup relationship to another sObject, Upsert Wizards skip "Step 2b", which means that the page rendering Step 2b is not marked as "Completed" and consequently Finish button remains grayed out in the final step. The fix is to mark the "Step 2b" page as "Completed" in its constructor instead of when it is rendered.