fix for W-13612705
W-13612705: [Data Loader v58 Regression] The rest of the field values are not updated if the field value of 1st data line in CSV file is blank when updating data via Data Loader Bulk API option

The regression was introduced in BulkLoadVisitor.java during the change made in PR #594 to optimize determining the header fields for each load batch. The fix is to revert the relevant code to the prior implementation.