Fix for W-14287051
Fix for W-14287051 - Unable to insert/update a file preview image into a rich text field after data loader version 58.0.3 and higher.

The issue is due to the code skipping the check for HTML tag if there are characters outside a HTML tag. Fix is to do the HTML tag check even if there are no chars outside a HTML tag.