improve readability of code for listing idlookup fields of parent
The code for listing idlookup fields of the parent object has 4-5 levels of nested for loops and if conditions with multiple predicates in each if statement. These are simplified by removing repeated if conditions to first determine that an sobject field is a lookup relationship field and it is createable or readable.