Fix for an issue with a whitespace char in install path
Fixed the following issue:
Data Loader creates a folder with "%20%" in the folder name for configs if installation path has a whitespace character in it. For example, if the installation path is /my/dir/with a space/v60.0.0/, data loader creates a folder /my/dir/with%20a%20space/v60.0.0/ and creates configs folder within it.