enable/disable undelete and hard delete buttions correctly
Undelete operation is possible only when SOAP API is selected. The Undelete button should be grayed out when Bulk or Bulk v2 API is selected in Settings dialog.

Similarly, Hard Delete operation is possible only when Bulk or Bulk v2 API is selected. The Hard Delete button should be grayed out when SOAP API is selected in Settings dialog.