Fix for issue #1244
REST Update fails if it includes update of a lookup relationship field using an idLookup field on the parent object (e.g. external id field).

The fix is to generate JSON string for the idLookup field setting in the parent object and set it to the relationship field in the child object.