ship an empty config.properties file
Shipped config.properties file is incomplete, in that it does not contain default settings for all properties. Moreover, it can cause unexpected side-effects if a setting value defers from its default setting value in the code. To fully make sure that the default settings are in effect as expected, the default config.properties file should be empty, i.e. no settings.